requirements.txt
Flask==2.2.2
google-cloud-datastore==2.8.2
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
six==1.16.0
python-jose==3.3.0